shared/pyproject.toml
[tool.poetry]
name = "pytegbot-shared"
version = "0.1.0"
description = "Shared models and config helpers for PyTegBot"
authors = ["k4m454k"]
packages = [{ include = "pytegbot_shared", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
pydantic = ">=2.10,<3.0"
pydantic-settings = ">=2.7,<3.0"
PyYAML = ">=6.0,<7.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
